refactor(ledger): collocate the transaction-count route registrars :hammer:
The Huma terminal registrar came out of count_handler.go and the Fiber
guard-chain registrars came out of routes.go; both now live in count_routes.go.

count_handler.go no longer references stdlib net/http once the Huma operation
constants leave with the registrar, so that import goes too.

No route, authz tuple, operation ID or path changed.
